+++ lib/kb_uploadmethods/kb_uploadmethodsClient.pm
@@ -1386,6 +1386,120 @@ report_ref is a string
  
 
 
+=head2 import_file_as_fba_model_from_staging
+
+  $returnVal = $obj->import_file_as_fba_model_from_staging($params)
+
+=over 4
+
+=item Parameter and return types
+
+=begin html
+
+<pre>
+$params is a kb_uploadmethods.FileToFBAModelParams
+$returnVal is a kb_uploadmethods.UploadMethodResult
+FileToFBAModelParams is a reference to a hash where the following keys are defined:
+	model_file has a value which is a string
+	compounds_file has a value which is a string
+	file_type has a value which is a string
+	genome has a value which is a string
+	biomass has a value which is a string
+	model_name has a value which is a string
+	workspace_name has a value which is a kb_uploadmethods.workspace_name
+workspace_name is a string
+UploadMethodResult is a reference to a hash where the following keys are defined:
+	obj_ref has a value which is a kb_uploadmethods.obj_ref
+	report_name has a value which is a kb_uploadmethods.report_name
+	report_ref has a value which is a kb_uploadmethods.report_ref
+obj_ref is a string
+report_name is a string
+report_ref is a string
+
+</pre>
+
+=end html
+
+=begin text
+
+$params is a kb_uploadmethods.FileToFBAModelParams
+$returnVal is a kb_uploadmethods.UploadMethodResult
+FileToFBAModelParams is a reference to a hash where the following keys are defined:
+	model_file has a value which is a string
+	compounds_file has a value which is a string
+	file_type has a value which is a string
+	genome has a value which is a string
+	biomass has a value which is a string
+	model_name has a value which is a string
+	workspace_name has a value which is a kb_uploadmethods.workspace_name
+workspace_name is a string
+UploadMethodResult is a reference to a hash where the following keys are defined:
+	obj_ref has a value which is a kb_uploadmethods.obj_ref
+	report_name has a value which is a kb_uploadmethods.report_name
+	report_ref has a value which is a kb_uploadmethods.report_ref
+obj_ref is a string
+report_name is a string
+report_ref is a string
+
+
+=end text
+
+=item Description
+
+
+
+=back
+
+=cut
+
+ sub import_file_as_fba_model_from_staging
+{
+    my($self, @args) = @_;
+
+# Authentication: required
+
+    if ((my $n = @args) != 1)
+    {
+	Bio::KBase::Exceptions::ArgumentValidationError->throw(error =>
+							       "Invalid argument count for function import_file_as_fba_model_from_staging (received $n, expecting 1)");
+    }
+    {
+	my($params) = @args;
+
+	my @_bad_arguments;
+        (ref($params) eq 'HASH') or push(@_bad_arguments, "Invalid type for argument 1 \"params\" (value was \"$params\")");
+        if (@_bad_arguments) {
+	    my $msg = "Invalid arguments passed to import_file_as_fba_model_from_staging:\n" . join("", map { "\t$_\n" } @_bad_arguments);
+	    Bio::KBase::Exceptions::ArgumentValidationError->throw(error => $msg,
+								   method_name => 'import_file_as_fba_model_from_staging');
+	}
+    }
+
+    my $url = $self->{url};
+    my $result = $self->{client}->call($url, $self->{headers}, {
+	    method => "kb_uploadmethods.import_file_as_fba_model_from_staging",
+	    params => \@args,
+    });
+    if ($result) {
+	if ($result->is_error) {
+	    Bio::KBase::Exceptions::JSONRPC->throw(error => $result->error_message,
+					       code => $result->content->{error}->{code},
+					       method_name => 'import_file_as_fba_model_from_staging',
+					       data => $result->content->{error}->{error} # JSON::RPC::ReturnObject only supports JSONRPC 1.1 or 1.O
+					      );
+	} else {
+	    return wantarray ? @{$result->result} : $result->result->[0];
+	}
+    } else {
+        Bio::KBase::Exceptions::HTTP->throw(error => "Error invoking method import_file_as_fba_model_from_staging",
+					    status_line => $self->{client}->status_line,
+					    method_name => 'import_file_as_fba_model_from_staging',
+				       );
+    }
+}
+ 
+
+
 =head2 import_tsv_as_expression_matrix_from_staging
 
   $returnVal = $obj->import_tsv_as_expression_matrix_from_staging($params)
@@ -3134,6 +3248,65 @@ workspace_name has a value which is a kb_uploadmethods.workspace_name
 a reference to a hash where the following keys are defined:
 staging_file_subdir_path has a value which is a string
 media_name has a value which is a string
+workspace_name has a value which is a kb_uploadmethods.workspace_name
+
+
+=end text
+
+=back
+
+
+
+=head2 FileToFBAModelParams
+
+=over 4
+
+
+
+=item Description
+
+required params:
+model_file: subdirectory file path for model file
+e.g.
+  for file: /data/bulk/user_name/file_name
+  staging_file_subdir_path is file_name
+  for file: /data/bulk/user_name/subdir_1/subdir_2/file_name
+  staging_file_subdir_path is subdir_1/subdir_2/file_name
+compounds_file: same as above for compound (only used for tsv)
+file_type: one of "tsv", "excel", "sbml"
+genome: the associated species genome
+biomasses: one or more biomass reactions in model
+model_name: output FBAModel object name
+workspace_name: workspace name/ID of the object
+
+
+=item Definition
+
+=begin html
+
+<pre>
+a reference to a hash where the following keys are defined:
+model_file has a value which is a string
+compounds_file has a value which is a string
+file_type has a value which is a string
+genome has a value which is a string
+biomass has a value which is a string
+model_name has a value which is a string
+workspace_name has a value which is a kb_uploadmethods.workspace_name
+
+</pre>
+
+=end html
+
+=begin text
+
+a reference to a hash where the following keys are defined:
+model_file has a value which is a string
+compounds_file has a value which is a string
+file_type has a value which is a string
+genome has a value which is a string
+biomass has a value which is a string
+model_name has a value which is a string
 workspace_name has a value which is a kb_uploadmethods.workspace_name