+++ lib/kb_uploadmethods/kb_uploadmethodsClient.py
@@ -334,6 +334,32 @@ class kb_uploadmethods(object):
             'kb_uploadmethods.import_tsv_or_excel_as_media_from_staging',
             [params], self._service_ver, context)
 
+    def import_file_as_fba_model_from_staging(self, params, context=None):
+        """
+        :param params: instance of type "FileToFBAModelParams" (required
+           params: model_file: subdirectory file path for model file e.g. for
+           file: /data/bulk/user_name/file_name staging_file_subdir_path is
+           file_name for file:
+           /data/bulk/user_name/subdir_1/subdir_2/file_name
+           staging_file_subdir_path is subdir_1/subdir_2/file_name
+           compounds_file: same as above for compound (only used for tsv)
+           file_type: one of "tsv", "excel", "sbml" genome: the associated
+           species genome biomasses: one or more biomass reactions in model
+           model_name: output FBAModel object name workspace_name: workspace
+           name/ID of the object) -> structure: parameter "model_file" of
+           String, parameter "compounds_file" of String, parameter
+           "file_type" of String, parameter "genome" of String, parameter
+           "biomass" of String, parameter "model_name" of String, parameter
+           "workspace_name" of type "workspace_name" (workspace name of the
+           object)
+        :returns: instance of type "UploadMethodResult" -> structure:
+           parameter "obj_ref" of type "obj_ref", parameter "report_name" of
+           type "report_name", parameter "report_ref" of type "report_ref"
+        """
+        return self._client.call_method(
+            'kb_uploadmethods.import_file_as_fba_model_from_staging',
+            [params], self._service_ver, context)
+
     def import_tsv_as_expression_matrix_from_staging(self, params, context=None):
         """
         :param params: instance of type "FileToMatrixParams" (required