+++ lib/kb_uploadmethods/kb_uploadmethodsServer.py
@@ -377,6 +377,10 @@ class Application(object):
                              name='kb_uploadmethods.import_tsv_or_excel_as_media_from_staging',
                              types=[dict])
         self.method_authentication['kb_uploadmethods.import_tsv_or_excel_as_media_from_staging'] = 'required'  # noqa
+        self.rpc_service.add(impl_kb_uploadmethods.import_file_as_fba_model_from_staging,
+                             name='kb_uploadmethods.import_file_as_fba_model_from_staging',
+                             types=[dict])
+        self.method_authentication['kb_uploadmethods.import_file_as_fba_model_from_staging'] = 'required'  # noqa
         self.rpc_service.add(impl_kb_uploadmethods.import_tsv_as_expression_matrix_from_staging,
                              name='kb_uploadmethods.import_tsv_as_expression_matrix_from_staging',
                              types=[dict])